data_analysis/requirements.txt
streamlit
langchain-core
langchain-ollama
langchain-openai
sqlalchemy
psycopg2-binary
pandas
matplotlib
seaborn
scikit-learn
plotly
python-dotenv